requirements.txt
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
sqlparse==0.4.4
psycopg2-binary==2.9.7
gunicorn==21.2.0